+++ src/createAMSClient.ts
@@ -15,6 +15,7 @@ interface AMSConfig {
 const createAMSClient = async (config: AMSConfig): Promise<FramedClient | FramedlessClient> => {
 
     (config as AMSConfig).debug && console.log(`[createAMSClient] ${config.framedMode? 'FramedClient': 'FramedlessClient'}`);
+    (config as AMSConfig).debug && console.time("createAMSClient");
 
     const logger = new AMSLogger(config.logger);
     const framedClientConfig = {
@@ -22,14 +23,13 @@ const createAMSClient = async (config: AMSConfig): Promise<FramedClient | Framed
     };
 
     const client = config.framedMode? new FramedClient(logger, framedClientConfig): new FramedlessClient(logger);
-
     await client.setup();
 
     (config as AMSConfig).debug && client.setDebug((config as AMSConfig).debug || false);
-
     GlobalConfiguration.debug = (config as AMSConfig).debug || false;
     GlobalConfiguration.silentError = (config as AMSConfig).silentError || true;
 
+    (config as AMSConfig).debug && console.timeEnd("createAMSClient");
     return client;
 }